first kernel

+++ main.cu
@@ -4,7 +4,7 @@
 
 __global__ void kernelA(int *A, int *x, int *b, int N){
   int tId = threadIdx.x + blockIdx.x * blockDim.x;
-  int mp = tId%1e4;
+  int mp = (int) tId%1e4;
   A[tId] = A[tId]*x[mp];
   atomicAdd(&b[mp],A[tId]);
 }
